requirements.txt
fastapi[standard]
uvicorn
python-multipart
requests
python-dotenv
python-telegram-bot